wip: start fixing wrong data

+++ src/encoder/encoder.h
@@ -8,9 +8,9 @@ struct aux {
   char **argv;
   int argc;
 
-  size_t data_length;
-  size_t encoded_data_length;
-  size_t compressed_data_length;
+  unsigned int data_length;
+  unsigned int encoded_data_length;
+  unsigned int compressed_data_length;
 
   char **dictionary;
   unsigned char *compressed_data;
